chore: move `try_file` to utils
