+++ mw2fcitx/utils.py
@@ -1,4 +1,9 @@
+import sys
+import re
 from copy import deepcopy
+from importlib import import_module
+import os
+import logging
 from typing import List
 from urllib3.util import Retry
 from requests import Session
@@ -6,6 +11,8 @@ from requests.adapters import HTTPAdapter
 
 
 from .version import PKG_VERSION
+
+log = logging.getLogger(__name__)
 
 
 def normalize(word):
@@ -72,3 +79,25 @@ def create_requests_session():
     s.mount('http://', HTTPAdapter(max_retries=retries))
     s.mount('https://', HTTPAdapter(max_retries=retries))
     return s
+
+
+def try_file(file):
+    log.debug(f"Finding config file: {file}")
+    if not os.access(file, os.R_OK):
+        log.error("File ({}) not readable.")
+        return False
+    file_realpath = os.path.realpath(file)
+    log.debug(f"Config file path: {file_realpath}")
+    file_path = os.path.dirname(file_realpath)
+    file_name = os.path.basename(file_realpath)
+    module_name = re.sub(r"\.py$", "", file_name)
+    config_file = False
+    try:
+        sys.path.insert(1, file_path)
+        config_file = import_module(module_name)
+    except Exception as e:
+        log.error(f"Error reading config: {str(e)}")
+        return False
+    finally:
+        sys.path.remove(file_path)
+    return config_file
